backend-python/requirements.txt
flask
pdfminer.six
google-generativeai
flask-cors